feat: add theme switcher script to App.webc
This commit adds a script to the head of App.webc that checks the user's local storage for a 'theme' item. If the item is 'light', it sets the data-theme attribute of the document to 'light'. If the item is 'dark', it sets the data-theme attribute to 'dark'. This allows for a user's theme preference to be saved and applied on subsequent visits.